+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - BREAKING CHANGE: Drop Python 3.5 and 3.6 support.
+- BREAKING CHANGE: Remove deprecated ``loop`` keyword argument for the SMTP class.
 - Change: The ``source_address`` argument now takes a (addr, port) tuple that is
   passed as the ``local_addr`` param to ``asyncio.create_connection``, allowing
   for binding to a specific IP. The new ``local_hostname`` argument that takes